api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.3",
    "fastapi[standard]>=0.123.8",
    "python-dotenv>=1.2.1",
]
